Preserve type of result arrays

+++ halo_tasks.py
@@ -59,7 +59,7 @@ def process_single_halo(mesh, data, halo_prop_list, a, z, cosmo,
         halo_result.update(halo_prop.calculate(cosmo, a, z, centre, halo_data))
 
     # Add the halo index to the result set
-    halo_result["index"] = (u.Quantity(index, unit=None), "Index of this halo in the input catalogue")
+    halo_result["index"] = (u.Quantity(index, unit=None, dtype=np.int64), "Index of this halo in the input catalogue")
 
     return halo_result
 
@@ -115,7 +115,7 @@ def process_halos(comm, data, mesh, halo_prop_list, a, z, cosmo,
     for halo_nr, result in enumerate(results):
         for name, (value, description) in result.items():
             if name not in result_arrays:
-                arr = astropy.units.Quantity(-np.ones(nr_halos, dtype=float), unit=value.unit)
+                arr = astropy.units.Quantity(-np.ones(nr_halos, dtype=value.dtype), unit=value.unit, dtype=value.dtype)
                 result_arrays[name] = (arr, description)
             result_arrays[name][0][halo_nr] = value
